requirements.txt
# Python dependencies for song identification
numpy
librosa
pyacoustid
chromaprint
torchaudio
torch
transformers